+++ packages/components/src/components/dropdown/dropdown.e2e.ts
@@ -6,13 +6,11 @@ import {
   findAll,
   getFocusedElementProp,
   skipAnimations,
-  waitForAnimationFrame,
 } from "../../tests/utils/puppeteer";
 import type { DropdownItem } from "../dropdown-item/dropdown-item";
 import type { Button } from "../button/button";
 import { mockConsole } from "../../tests/utils/logging";
 import { CSS as DROPDOWN_ITEM_CSS } from "../dropdown-item/resources";
-import { CSS } from "./resources";
 
 mockConsole();
 
@@ -475,207 +473,6 @@ it("should focus the first item on open (multi) when items are selected", async 
           .ariaActiveDescendantElement?.id,
     ),
   ).toEqual("item-1");
-});
-
-describe("scrolling", () => {
-  it("focused item should be in view when long", async () => {
-    const page = await newE2EPage();
-
-    await page.setContent(
-      html`<calcite-dropdown>
-        <calcite-button slot="trigger">Open Dropdown</calcite-button>
-        <calcite-dropdown-group>
-          <calcite-dropdown-item id="item-1">1</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-2">2</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-3">3</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-4">4</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-5">5</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-6">6</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-7">7</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-8">8</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-9">9</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-10">10</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-11">11</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-12">12</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-13">13</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-14">14</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-15">15</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-16">16</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-17">17</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-18">18</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-19">19</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-20">20</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-21">21</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-22">22</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-23">23</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-24">24</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-25">25</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-26">26</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-27">27</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-28">28</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-29">29</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-30">30</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-41">41</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-42">42</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-43">43</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-44">44</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-45">45</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-46">46</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-47">47</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-48">48</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-49">49</calcite-dropdown-item>
-          <calcite-dropdown-item id="item-50">50</calcite-dropdown-item>
-        </calcite-dropdown-group>
-      </calcite-dropdown>`,
-    );
-    await page.waitForChanges();
-
-    const element = await page.find("calcite-dropdown");
-    await element.callMethod("setFocus");
-    await page.waitForChanges();
-
-    const dropdownOpenEventSpy = await page.spyOnEvent("calciteDropdownOpen");
-    await page.keyboard.press("ArrowUp");
-    await page.waitForChanges();
-    await dropdownOpenEventSpy.next();
-
-    expect(
-      await page.evaluate(
-        () =>
-          document.querySelector("calcite-dropdown")!.shadowRoot!.querySelector("slot[name='trigger']")!
-            .ariaActiveDescendantElement?.id,
-      ),
-    ).toEqual("item-50");
-
-    const item = await page.find("#item-50");
-
-    expect(await item.isIntersectingViewport()).toBe(true);
-  });
-
-  describe("max-items", () => {
-    const maxItems = 7;
-
-    it("control max items displayed", async () => {
-      const page = await newE2EPage();
-      await page.setContent(
-        html` <calcite-dropdown max-items="${maxItems}">
-          <calcite-button slot="trigger">Open Dropdown</calcite-button>
-          <calcite-dropdown-group group-title="First group">
-            <calcite-dropdown-item id="item-1">1</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-2">2</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-3">3</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-4">4</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-5">5</calcite-dropdown-item>
-          </calcite-dropdown-group>
-          <calcite-dropdown-group group-title="Second group">
-            <calcite-dropdown-item id="item-6">6</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-7">7</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-8">8</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-9">9</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-10">10</calcite-dropdown-item>
-          </calcite-dropdown-group>
-        </calcite-dropdown>`,
-      );
-
-      const element = await page.find("calcite-dropdown");
-      const dropdownOpenEventSpy = await page.spyOnEvent("calciteDropdownOpen");
-      await element.click();
-      await dropdownOpenEventSpy.next();
-
-      const items = await findAll(page, "calcite-dropdown-item");
-
-      for (let i = 0; i < items.length; i++) {
-        expect(await items[i].isIntersectingViewport()).toBe(i <= maxItems - 1);
-      }
-
-      const newMaxItems = 4;
-      element.setProperty("maxItems", newMaxItems);
-      await page.waitForChanges();
-
-      for (let i = 0; i < items.length; i++) {
-        expect(await items[i].isIntersectingViewport()).toBe(i <= newMaxItems - 1);
-      }
-
-      const totalItems = 10;
-      element.setProperty("maxItems", totalItems);
-      await page.waitForChanges();
-
-      for (let i = 0; i < items.length; i++) {
-        expect(await items[i].isIntersectingViewport()).toBe(true);
-      }
-
-      // no scroller should be present when max-items === items
-      const scroller = await page.find(`calcite-dropdown >>> .${CSS.content}`);
-      expect(await scroller.getProperty("scrollHeight")).toBe(await scroller.getProperty("clientHeight"));
-    });
-
-    it("does not display a scrollbar on subsequent opens when max-items matches the number of items", async () => {
-      const page = await newE2EPage();
-      // the number of items must match `max-items` for this scenario
-      const items = Array.from({ length: maxItems }, (_, index) => index + 1)
-        .map((position) => html`<calcite-dropdown-item id="item-${position}">${position}</calcite-dropdown-item>`)
-        .join("");
-
-      await page.setContent(
-        html` <calcite-dropdown max-items="${maxItems}">
-          <calcite-button slot="trigger">Open Dropdown</calcite-button>
-          <calcite-dropdown-group group-title="First group">${items}</calcite-dropdown-group>
-        </calcite-dropdown>`,
-      );
-
-      const element = await page.find("calcite-dropdown");
-      const scroller = await page.find(`calcite-dropdown >>> .${CSS.content}`);
-      const dropdownOpenEventSpy = await page.spyOnEvent("calciteDropdownOpen");
-      const dropdownCloseEventSpy = await page.spyOnEvent("calciteDropdownClose");
-
-      async function assertNoScrollbar(): Promise<void> {
-        await waitForAnimationFrame(page);
-        expect(await scroller.getProperty("scrollHeight")).toBe(await scroller.getProperty("clientHeight"));
-      }
-
-      await element.click();
-      await dropdownOpenEventSpy.next();
-      await assertNoScrollbar();
-
-      await element.click();
-      await dropdownCloseEventSpy.next();
-
-      await element.click();
-      await dropdownOpenEventSpy.next();
-      await assertNoScrollbar();
-    });
-
-    it("does not scroll to selected item on open when max-items causes selected item to be beyond scroller", async () => {
-      const page = await newE2EPage();
-      await page.setContent(
-        html` <calcite-dropdown max-items="${maxItems}">
-          <calcite-button slot="trigger">Open Dropdown</calcite-button>
-          <calcite-dropdown-group group-title="First group">
-            <calcite-dropdown-item id="item-1">1</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-2">2</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-3">3</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-4">4</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-5">5</calcite-dropdown-item>
-          </calcite-dropdown-group>
-          <calcite-dropdown-group group-title="Second group">
-            <calcite-dropdown-item id="item-6">6</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-7">7</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-8">8</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-9">9</calcite-dropdown-item>
-            <calcite-dropdown-item id="item-10" selected>10</calcite-dropdown-item>
-          </calcite-dropdown-group>
-        </calcite-dropdown>`,
-      );
-      const selectedItem = await page.find("#item-10");
-      const element = await page.find("calcite-dropdown");
-      const dropdownOpenEventSpy = await page.spyOnEvent("calciteDropdownOpen");
-
-      await element.click();
-      await dropdownOpenEventSpy.next();
-
-      expect(await selectedItem.isIntersectingViewport()).toBe(false);
-    });
-  });
 });
 
 it("closes when a selection is made", async () => {
